+++ enterprise/internal/codeintel/bundles/database/gen.go
@@ -1,0 +1,4 @@
+package database
+
+//go:generate env GOBIN=$PWD/.bin GO111MODULE=on go install github.com/efritz/go-mockgen
+//go:generate $PWD/.bin/go-mockgen -f github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/bundles/database -i Database -o mock_database_test.go
