+++ enterprise/internal/codeintel/bundles/database/ranges.go
