+++ enterprise/internal/codeintel/bundles/database/ranges_test.go
